Search page fixes
Fixed some elements of the search page. Radio options now function as they are supposed to. Removed "Owner" option as the objects are not linked. added search bar for criteria text. "Search" button no longer sends you to whitelabel error page, however there is currently no page for it to send you to so it sends you back to the list of pets