Starting Page + Search Progress
I added an  "index" page, it is a simple page that just has two buttons, one to take you to the pets list and one to take you to the owners list. I tried to work on the search feature but the guide that the teacher herself gave me when I asked for help did not work. I have managed to get it to the point where it will take you to the correct area and all of the links work, but nothing will show up when you actually try to filter anything.